Basic structure of Block class
